updated to work properly, added commenting

+++ aws-s3/scripts-ccao-data-raw-us-east-1/spatial-access.R
@@ -44,30 +44,45 @@ lapply(api_info, pull_and_write)
 
 ##### BUILDING FOOTPRINTS #####
 
-temp_file <- tempfile(fileext = ".geojson")
-
 remote_file <- file.path(
-  AWS_S3_RAW_BUCKET, "spatial", "access",
+  AWS_S3_RAW_BUCKET, "spatial", "other",
   "building_footprint", paste0(strftime(Sys.Date(), "%Y"), ".geojson"))
 
-footprints <- opq("Cook County United States") %>%
-  add_osm_feature(key = "building", value = "geometry") %>%
-  osmdata_sf()
+if (!aws.s3::object_exists(remote_file)) {
 
-footprints <- data.frame(footprints[[1]]) %>%
-  st_as_sf(cords = c("X1", "X2"), crs = 4326) %>%
-  pull(geometry) %>%
-  st_make_valid() %>%
-  st_intersection(
-    st_read(
-      "https://opendata.arcgis.com/datasets/ea127f9e96b74677892722069c984198_1.geojson"
-    )
-  ) %>%
-  pull(geometry) %>%
-  st_write(temp_file)
+  temp_file <- tempfile(fileext = ".geojson")
 
-aws.s3::put_object(temp_file, remote_file)
-file.remove(temp_file)
+  # Gather building footprints from OSM
+  footprints <- opq("Cook County United States") %>%
+    add_osm_feature(key = "building") %>%
+    osmdata_sf()
+
+  # Only need polygons
+  footprints <- footprints$osm_polygons
+
+  # Transform
+  footprints %>%
+    st_transform(crs = 3435) %>%
+    st_make_valid() %>%
+
+    # Intersect with Cook County border
+    st_intersection(
+      st_transform(
+        st_read(
+          "https://opendata.arcgis.com/datasets/ea127f9e96b74677892722069c984198_1.geojson"
+          ),
+        3435
+        )
+      ) %>%
+    pull(geometry) %>%
+
+    # Upload to AWS
+    st_write(temp_file)
+
+  aws.s3::put_object(temp_file, remote_file, multipart = TRUE)
+  file.remove(temp_file)
+
+}
 
 # Cleanup
 rm(list = ls())
